band 95-100: six functions EXACT; two more unsigned domains; 64 stale markers
Six functions to 100%, each from a named source bug rather than a re-spelling:

  0x00118bf0  HeapStats                        97.34 -> 100  the three-field
              zero of `_HEAPINFO` is a `memset`, which clears with a FRESH zero
              register; three field stores reuse an existing one.
  0x00177480  CImagePaletteNode::LoadBmpFile   98.72 -> 100  the second argument
              of the trailing Build() is the function's own `arg` parameter
              (retail reloads it from [esp+0x864]); we passed a literal 0.
  0x00175d50  CRezImage::Fill                  96.x  -> 100  `value &= 0xff` is a
              memory read-modify-write; retail load/and/stores, i.e. a NEW local
              (`i32 fill = value & 0xff`) coalesced onto the parameter's slot.
  0x00152f20  CImage::Resolve                  98.89 -> 100  \
  0x00153380  CImage::Reload                   99.40 -> 100   | RezTypeTag is
  0x00163e50  CDDrawSurfacePair::LoadImageA    98.91 -> 100  /  UNSIGNED
  0x0013a0f0  CRezDirNode::Load                98.97 -> 100  CParserObjList::
              m_count is unsigned.

The two unsigned domains are the same mechanism the AnimWorkerAct sweep found:
`--branches --diff` reports `jg`/`jge` where retail has `ja`/`jae`, everything
else byte-identical. RezTypeTag gets GZ_ENUM_BEGIN_SPLIT(u32) +
GZ_ENUM_RETURN(RezTypeTag, u32) on CParseSource::GetEntryTag; WorldInitError
gets the same treatment plus GZ_ENUM_STORAGE on CDDrawSurfaceMgr::m_lastError
(CGruntzMgr::ReportWorldStatus 98.87 -> 99.98).

Also:
  0x00022be0  CCheatMgr::AddCheat              97.76 -> 98.57  the duplicate
              check is `Lookup(...) ? existing : NULL`, which drops the
              AddrWord<char> pun; residue is one materialisation.
  0x001495d0  CDDrawShadeBlit::EncodeRle16     the RLE run length is unsigned
              (`jbe`); all 11 branches now agree. Current % dips 79.2 -> 77.5 -
              MAX holds; the real blocker is a frame one dword too big.

Housekeeping: 64 stale `// @early-stop` markers removed from functions that are
now EXACT, so `rg '@early-stop' src` is a true worklist again.

+++ src/Image/ImageRle16Encode.cpp
@@ -59,7 +59,7 @@ void* CDDrawShadeBlit::EncodeRle16(const u8* src) {
                     x2 += static_cast<i32>(m_rleData[srcidx]) - 0x80;
                     srcidx++;
                 } else {
-                    i32 n = src[srcidx];
+                    u32 n = src[srcidx];
                     outidx++;
                     if (n > 0) {
                         const u8* run = src + srcidx + 1;
@@ -70,7 +70,7 @@ void* CDDrawShadeBlit::EncodeRle16(const u8* src) {
                             out[outidx + 1] = static_cast<u8>((px >> 8));
                             outidx += 2;
                             k++;
-                        } while (k < n);
+                        } while (k < static_cast<i32>(n));
                     }
                     x2 += static_cast<i32>(m_rleData[srcidx]);
                     srcidx += static_cast<i32>(m_rleData[srcidx]) + 1;
